Calculate readiness for tasks without checks based on start time.
This closes a readiness calculation gap where brand new tasks from an
initially deployed application and long-running tasks part of a
rolling-upgrade (with all new tasks of the deployment having crashed
momentarily and not be included in the API response) could not be
distinguished properly if readiness check results were also missing for
a task in question.

We solve the problem by looking at the age of a task: If its start time
is still within the timeout period of the configured readiness check
probe, we consider the task to be a new one and mark it as unready.
Otherwise, we assume the task to be a long-running one and mark it as
ready.

The new approach supersedes the one based on the number of task versions
available.